Added EchelonForm. Moved SafeDivision GE there. API for simple GE added
